pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "openai>=1.76.2",
    "pillow>=11.2.1",
    "streamlit>=1.45.0",
    "zipfile36>=0.1.3",
]
